+++ src/features/vacancy/search/ui/index.tsx
@@ -35,7 +35,6 @@ export const SearchVacancies = ({
     filters: filterResult?.filters,
   });
 
-  console.log(searchResult);
   return (
     <div className={cls("search-vacancies", className)}>
       <SearchBox